Replaced stringstream stod and stoi implementation with a wrapper around strtod and strtol. It checks for the same conditions as stod and stoi, but provides the same interface as strtod and strtol.
Using that interface, I was also able to remove all calls to string::substr
within datatable.cpp. This could drastically speed up loading
as it avoids copying parts of the string over and over again.

Instead, it just uses a bit of pointers to move ahead on the original string,
meaning no string copying is done.