Add explicit const char * constructors to avoid ambiguity
Approximants that has a constructor that takes a single DataTable
as argument could not decide what constructor to use when a
const char * was passed, as DataTable has a const char * constructor.

We avoid this ambiguity by adding an explicit const char * constructor
to all Approximants, regardless of whether the ambiguity currently
exists or not.