Allow for multiple outputs for eval in the Python and C interfaces
Only tested with one-dimensional output, as a bug in the C++ back-end
prevents from testing for multi-dimensional outputs.

If a single point is evaluated, the function currently returns a
scalar for the one-dimensional output case, and a list of the results
for the multi-dimensional output case. This is subject to change.

NOTE: Most, if not all, Python examples are probably broken by this
commit.